Remove Integer and Decimal value and leave only Number value

+++ KrasiMilan/QL/src/main/java/nl/uva/softwcons/ast/expression/literal/IntegerLiteral.java
@@ -1,21 +1,19 @@
 package nl.uva.softwcons.ast.expression.literal;
 
-import java.math.BigInteger;
-
 import nl.uva.softwcons.ast.LineInfo;
 import nl.uva.softwcons.ast.expression.Expression;
 import nl.uva.softwcons.ast.expression.ExpressionVisitor;
 
 public class IntegerLiteral extends Expression {
-    private BigInteger value;
+    private int value;
     private LineInfo lineInfo;
 
     public IntegerLiteral(int value, LineInfo lineInfo) {
-        this.value = BigInteger.valueOf(value);
+        this.value = value;
         this.lineInfo = lineInfo;
     }
 
-    public BigInteger getValue() {
+    public int getValue() {
         return value;
     }
